@@ -1,16 +1,20 @@
 module github.com/ysoftman/watchDust
 
-go 1.22
+go 1.23.0
+
+toolchain go1.24.1
 
 require (
-	github.com/BurntSushi/toml v0.3.1
-	github.com/PuerkitoBio/goquery v1.5.1
+	github.com/BurntSushi/toml v1.5.0
+	github.com/PuerkitoBio/goquery v1.10.2
+	github.com/caarlos0/env/v11 v11.3.1
 	github.com/robfig/cron v1.2.0
-	google.golang.org/appengine/v2 v2.0.1
+	google.golang.org/appengine/v2 v2.0.6
 )
 
 require (
-	github.com/andybalholm/cascadia v1.1.0 // indirect
-	github.com/golang/protobuf v1.3.3 // indirect
-	golang.org/x/net v0.0.0-20200202094626-16171245cfb2 // indirect
+	github.com/andybalholm/cascadia v1.3.3 // indirect
+	github.com/golang/protobuf v1.5.4 // indirect
+	golang.org/x/net v0.38.0 // indirect
+	google.golang.org/protobuf v1.36.6 // indirect
 )
